refactor(terminal): consolidate MachineRef onto the merged config-model type
Config foundation PR #1888 merged into pu/terminal and now owns the
canonical MachineRef (apps/web/src/lib/repositories/page-agent-repository.ts).
Deletes this PR's local duplicate (machine-ref.ts) and imports the type
from there instead — exactly one MachineRef declaration now exists repo-wide.

machineRefId/machineRefEquals/machineRefFromId (tools-layer helpers, not
part of the repository's canonical contract) move into sandbox-tools.ts
as local functions, importing only the MachineRef type from the repository
module (import type — no DB import, no runtime footprint, so the pure
factory's "imports no DB" contract still holds).

Stale "sibling Config model PR" comments updated to reflect that the
config foundation is merged; wiring PageAgentConfig.machines into
ToolExecutionContext per-turn remains the sibling "route tools to the
active machine" PR's job.
